Added create_veteran to helper module and used in multi_correspondences
