+++ db/seeds/helpers/seed_helpers.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module SeedHelpers
+  def create_veteran(options = {})
+    @file_number += 1
+    @participant_id += 1
+    params = {
+      file_number: format("%<n>09d", n: @file_number),
+      participant_id: format("%<n>09d", n: @participant_id)
+    }
+    create(:veteran, params.merge(options))
+  end
+end
